requirements.txt
async-generator==1.10
attrs==21.4.0
Brotli==1.0.9
cachetools==5.0.0
certifi==2021.10.8
idna==3.3
mutagen==1.45.1
outcome==1.1.0
pycryptodomex==3.14.1
python-mpd2==3.0.5
re-ircbot==1.5.5
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
websockets==10.3
yt-dlp==2022.4.8
